+++ crates/vm/levm/src/execution_handlers.rs
@@ -9,19 +9,16 @@ use bytes::Bytes;
 
 impl<'a> VM<'a> {
     pub fn handle_precompile_result(
-        &mut self,
         precompile_result: Result<Bytes, VMError>,
+        gas_limit: u64,
+        gas_remaining: u64,
     ) -> Result<ContextResult, VMError> {
         match precompile_result {
             Ok(output) => Ok(ContextResult {
                 result: TxResult::Success,
-                gas_used: {
-                    let callframe = self.current_call_frame()?;
-                    callframe
-                        .gas_limit
-                        .checked_sub(callframe.gas_remaining)
-                        .ok_or(InternalError::Underflow)?
-                },
+                gas_used: gas_limit
+                    .checked_sub(gas_remaining)
+                    .ok_or(InternalError::Underflow)?,
                 output,
             }),
             Err(error) => {
@@ -31,7 +28,7 @@ impl<'a> VM<'a> {
 
                 Ok(ContextResult {
                     result: TxResult::Revert(error),
-                    gas_used: self.current_call_frame()?.gas_limit,
+                    gas_used: gas_limit,
                     output: Bytes::new(),
                 })
             }