+++ crates/vm/levm/src/vm.rs
@@ -16,7 +16,7 @@ use crate::{
 };
 use bytes::Bytes;
 use ethrex_common::{
-    Address, H256, U256,
+    Address, H160, H256, U256,
     tracing::CallType,
     types::{Log, Transaction},
 };
@@ -28,7 +28,7 @@ use std::{
 
 pub type Storage = HashMap<U256, H256>;
 
-#[derive(Debug, Clone, Default)]
+#[derive(Debug, Clone, Copy, Default)]
 pub enum VMType {
     #[default]
     L1,
@@ -180,7 +180,16 @@ impl<'a> VM<'a> {
     /// Main execution loop.
     pub fn run_execution(&mut self) -> Result<ContextResult, VMError> {
         if self.is_precompile(&self.current_call_frame()?.to) {
-            return self.execute_precompile();
+            let vm_type = self.vm_type;
+            let call_frame = self.current_call_frame_mut()?;
+
+            return Self::execute_precompile(
+                vm_type,
+                call_frame.code_address,
+                &call_frame.calldata,
+                call_frame.gas_limit,
+                &mut call_frame.gas_remaining,
+            );
         }
 
         loop {
@@ -213,27 +222,23 @@ impl<'a> VM<'a> {
     }
 
     /// Executes precompile and handles the output that it returns, generating a report.
-    pub fn execute_precompile(&mut self) -> Result<ContextResult, VMError> {
-        let vm_type = self.vm_type.clone();
-
-        let callframe = self.current_call_frame_mut()?;
-
-        let precompile_result = match vm_type {
-            VMType::L1 => precompiles::execute_precompile(
-                callframe.code_address,
-                &callframe.calldata,
-                &mut callframe.gas_remaining,
-            ),
-            VMType::L2 => l2_precompiles::execute_precompile(
-                callframe.code_address,
-                &callframe.calldata,
-                &mut callframe.gas_remaining,
-            ),
+    pub fn execute_precompile(
+        vm_type: VMType,
+        code_address: H160,
+        calldata: &Bytes,
+        gas_limit: u64,
+        gas_remaining: &mut u64,
+    ) -> Result<ContextResult, VMError> {
+        let execute_precompile = match vm_type {
+            VMType::L1 => precompiles::execute_precompile,
+            VMType::L2 => l2_precompiles::execute_precompile,
         };
 
-        let ctx_result = self.handle_precompile_result(precompile_result)?;
-
-        Ok(ctx_result)
+        Self::handle_precompile_result(
+            execute_precompile(code_address, calldata, gas_remaining),
+            gas_limit,
+            *gas_remaining,
+        )
     }
 
     /// True if external transaction is a contract creation