+++ crates/vm/tracing.rs
@@ -28,14 +28,9 @@ impl Evm {
             Evm::REVM { state } => {
                 REVM::trace_tx_calls(&block.header, tx, state, only_top_call, with_log)
             }
-            Evm::LEVM { db, vm_type } => LEVM::trace_tx_calls(
-                db,
-                &block.header,
-                tx,
-                only_top_call,
-                with_log,
-                vm_type.clone(),
-            ),
+            Evm::LEVM { db, vm_type } => {
+                LEVM::trace_tx_calls(db, &block.header, tx, only_top_call, with_log, *vm_type)
+            }
         }
     }
 
@@ -50,7 +45,7 @@ impl Evm {
     ) -> Result<(), EvmError> {
         match self {
             Evm::REVM { state } => REVM::rerun_block(block, state, stop_index),
-            Evm::LEVM { db, vm_type } => LEVM::rerun_block(db, block, stop_index, vm_type.clone()),
+            Evm::LEVM { db, vm_type } => LEVM::rerun_block(db, block, stop_index, *vm_type),
         }
     }
 }